server/requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
certifi==2025.4.26
distlib==0.3.9
django==5.0.6
django-cors-headers==4.3.1
django-ninja==1.1.0
filelock==3.18.0
packaging==25.0
platformdirs==4.3.7
pydantic==2.7.2
pydantic-core==2.18.3
pyjwt==2.6.0
setuptools==80.1.0
sqlparse==0.5.0
typing-extensions==4.12.1
virtualenv==20.30.0
